Extract terminal response filtering from app.py into app_terminal_response.py
